fitness/requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
black==22.1.0
click==8.0.4
Django==4.0.3
django-bootstrap-v5==1.0.11
django-extensions==3.1.5
django-reversion==5.0.0
django-simple-history==3.0.0
gunicorn==20.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
prettytable==3.2.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tomli==2.0.1
wcwidth==0.2.5
Werkzeug==2.1.0
whitenoise==6.0.0
django-bootstrap-breadcrumbs
